requirements.txt
django>=2.1.7
configobj>=5.0.6
path.py>=11.5.0
psycopg2>=2.7.7
